+++ Android/MMKV/mmkv/src/main/cpp/MMKV.h
@@ -38,7 +38,8 @@ class AESCrypt;
 enum MMKVMode : uint32_t {
     MMKV_SINGLE_PROCESS = 0x1,
     MMKV_MULTI_PROCESS = 0x2,
-    MMKV_ASHMEM = 0x4,
+    CONTEXT_MODE_MULTI_PROCESS = 0x4, // in case someone mistakenly pass Context.MODE_MULTI_PROCESS
+    MMKV_ASHMEM = 0x8,
 };
 
 class MMKV {
@@ -191,7 +192,9 @@ public:
 
     bool getVectorForKey(const std::string &key, std::vector<std::string> &result);
 
-    size_t getValueSizeForKey(const std::string &key);
+    size_t getValueSizeForKey(const std::string &key, bool acutalSize);
+
+    int32_t writeValueToBuffer(const std::string &key, void *ptr, int32_t size);
 
     bool containsKey(const std::string &key);