+++ Android/MMKV/mmkv/src/main/cpp/native-bridge.cpp
@@ -160,7 +160,7 @@ static JNIEnv *getCurrentEnv() {
 
 MMKVRecoverStrategic onMMKVCRCCheckFail(const std::string &mmapID) {
     auto currentEnv = getCurrentEnv();
-    if (g_currentJVM && g_callbackOnCRCFailID) {
+    if (currentEnv && g_callbackOnCRCFailID) {
         jstring str = string2jstring(currentEnv, mmapID);
         auto strategic = currentEnv->CallStaticIntMethod(g_cls, g_callbackOnCRCFailID, str);
         return static_cast<MMKVRecoverStrategic>(strategic);
@@ -701,14 +701,12 @@ extern "C" JNIEXPORT void JNICALL Java_com_tencent_mmkv_MMKV_close(JNIEnv *env, 
     }
 }
 
-extern "C" JNIEXPORT jint JNICALL Java_com_tencent_mmkv_MMKV_valueSize(JNIEnv *env,
-                                                                       jobject instance,
-                                                                       jlong handle,
-                                                                       jstring oKey) {
+extern "C" JNIEXPORT jint JNICALL Java_com_tencent_mmkv_MMKV_valueSize(
+    JNIEnv *env, jobject instance, jlong handle, jstring oKey, jboolean actualSize) {
     MMKV *kv = reinterpret_cast<MMKV *>(handle);
     if (kv && oKey) {
         string key = jstring2string(env, oKey);
-        return kv->getValueSizeForKey(key);
+        return static_cast<jint>(kv->getValueSizeForKey(key, (bool) actualSize));
     }
     return 0;
 }
@@ -723,4 +721,32 @@ extern "C" JNIEXPORT void JNICALL Java_com_tencent_mmkv_MMKV_setLogReDirecting(J
                                                                                jclass type,
                                                                                jboolean enable) {
     g_isLogRedirecting = (enable == JNI_TRUE);
+}
+
+extern "C" JNIEXPORT jlong JNICALL Java_com_tencent_mmkv_MMKV_createNB(JNIEnv *env,
+                                                                       jobject instance,
+                                                                       jint size) {
+    auto ptr = malloc(static_cast<size_t>(size));
+    if (!ptr) {
+        MMKVError("fail to create NativeBuffer:%s", strerror(errno));
+        return 0;
+    }
+    return reinterpret_cast<jlong>(ptr);
+}
+
+extern "C" JNIEXPORT void JNICALL Java_com_tencent_mmkv_MMKV_destroyNB(JNIEnv *env,
+                                                                       jobject instance,
+                                                                       jlong pointer,
+                                                                       jint size) {
+    free(reinterpret_cast<void *>(pointer));
+}
+
+extern "C" JNIEXPORT jint JNICALL Java_com_tencent_mmkv_MMKV_writeValueToNB(
+    JNIEnv *env, jobject instance, jlong handle, jstring oKey, jlong pointer, jint size) {
+    MMKV *kv = reinterpret_cast<MMKV *>(handle);
+    if (kv && oKey) {
+        string key = jstring2string(env, oKey);
+        return kv->writeValueToBuffer(key, reinterpret_cast<void *>(pointer), size);
+    }
+    return -1;
 }